Add files via upload
As I said before, please remember to place it into your standalone folder, and change it to your own logo! This is the logo made for DantoxRP.
And if you want it to display the date as well, then please go to my profile and find the Dato folder as well.